Seperated menu from main and helper functions
